refactor(quests): extract chain-free XP-queue settlement out of onchain-queue
The quest paths (/api/quests/daily, quest-evaluation and every action route
that schedules it) only needed the DB-only settlement helpers, but importing
them from lib/solana/onchain-queue dragged @solana/web3.js + @coral-xyz/anchor
into those graphs — a real dev-compile and cold-start cost on routes that never
touch the chain.

Move retryQuestXpForUser / creditQuestXpRows / creditXpAndSettle / bumpRetry
verbatim into lib/gamification/xp-queue-settlement.ts (server-only marker plus
type-only imports, all clients passed as parameters), repoint the importers with
no compat re-export, and add a no-restricted-imports guard so lib/gamification
can never reach back into @/lib/solana/*.

Behaviour-neutral: src/lib/solana/__tests__/onchain-queue.test.ts passes
unmodified.
